move decodeManifest and decodeResources to the ResourceDecoder
